extract mass-rollout controller
