+++ config/routes.rb
@@ -83,12 +83,11 @@ Samson::Application.routes.draw do
   resources :commands, except: [:edit]
 
   resources :deploy_groups do
-    member do
-      post :deploy_all
-      get :create_all_stages_preview
-      post :create_all_stages
-      post :merge_all_stages
-      post :delete_all_stages
+    resource :mass_rollouts, only: [:new, :create, :destroy] do
+      collection do
+        post :merge
+        post :deploy
+      end
     end
   end
